requirements.txt
django
drf_spectacular
django-cors-headers
djangorestframework

openai
schedule
llama_index
python-dotenv
psycopg2-binary

pydantic
langchain
langchain-core
langchain-openai
